fixed udp_stub test filename
udp_stub_connection.go was causing an issue when flag was used, since it
was being included in the finished build, the cmdline arguments for the
testing package appeared in the usage instructions.

simply adding _test to the end of the file name prevents it from being
included in a normal build, along with the testing package.